[Hosts|Policies|Collect] Merge SoSHost into Policy, drop SoSHost
Moves the functions from `SoSHost` into `Policy`, and adjusts the calls
in `SoSNode` appropriately.

Adds a `pkg_version` method to `PackageManager` to satisfy the version
checking requirements that had previously been using `SoSHost`.

Finally drops the sos/collector/hosts directory from the tree as it is
no longer used.

At this point, `sos collect` should be able to be run on any
distribution that has a defined `Policy`. Note that new policies will
need to specify 2 new class attrs, `sos_pkg_name` and `sos_bin_path` in
order for them to be supported as well.
